# lib/api-zod/src/generated/api/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Accepts a base64-encoded image and returns an AI-generated description and tags
 * @summary Analyze a photo with AI
 */
export const AnalyzePhotoBody = zod.object({
  imageBase64: zod.string().describe("Base64-encoded JPEG image data"),
  photoId: zod.string().describe("Unique identifier for the photo"),
});

export const AnalyzePhotoResponse = zod.object({
  photoId: zod.string(),
  description: zod.string().describe("A 1-2 sentence description of the photo"),
  tags: zod.array(zod.string()).describe("Array of descriptive keywords"),
  sceneType: zod.string().describe("Scene type: indoor, outdoor, or mixed"),
  timeOfDay: zod
    .string()
    .describe("Time of day: day, night, dawn, dusk, or unknown"),
});

/**
 * Given a search query and list of indexed photo descriptions, returns ranked matches
 * @summary Search photos by natural language query
 */
export const SearchPhotosBody = zod.object({
  query: zod.string().describe("Natural language search query"),
  photos: zod
    .array(
      zod.object({
        id: zod.string(),
        description: zod.string(),
        tags: zod.array(zod.string()),
        sceneType: zod.string(),
        timeOfDay: zod.string(),
      }),
    )
    .describe("List of indexed photo descriptions to search through"),
});

export const SearchPhotosResponse = zod.object({
  results: zod.array(
    zod.object({
      id: zod.string(),
      relevanceScore: zod.number().describe("Score between 0 and 1"),
      matchReason: zod
        .string()
        .describe("Brief explanation of why this photo matches"),
    }),
  ),
});